Adds a data row object
This will help facilitate the loading of entities.

I will note, there is an issue here where the Storage interfaces know of the Row class, however, the row class does not belong in the 1st level.
While this violates our architecture (slightly, rows are fairly stable) it will be here for now.
In the future we can create a builder class that our storage interacts with.
The builder would take the "load from" functionality of the storages and simply pass them completed lists of entities. This means that at the top level, there is no indication of building entities.