Implements testing for JSON with nested arrays
In order to keep the project on track, we are leaving in place the more or less hard coded requirements of our JSON data being arrays of objects, each of which contain a type and only primitives or arrays of primitives.

This will be better realized once we have our entity ID's implemented since this eliminates the concern for a lot of our nested objects.

Better exception handling could be used when accessing values of our JSON objects.